+++ packages/authhero/src/management-app.ts
@@ -5,7 +5,7 @@ import { addDataHooks } from "./hooks";
 // import { domainRoutes } from "./routes/management-api/domains";
 import { userRoutes } from "./routes/management-api/users";
 // import { keyRoutes } from "./routes/management-api/keys";
-// import { usersByEmailRoutes } from "./routes/management-api/users-by-email";
+import { usersByEmailRoutes } from "./routes/management-api/users-by-email";
 import { clientRoutes } from "./routes/management-api/clients";
 import { tenantRoutes } from "./routes/management-api/tenants";
 import { logRoutes } from "./routes/management-api/logs";
@@ -34,7 +34,7 @@ export default function create(params: CreateAuthParams) {
     // .route("/api/v2/domains", domainRoutes)
     .route("/api/v2/users", userRoutes)
     // .route("/api/v2/keys/signing", keyRoutes)
-    // .route("/api/v2/users-by-email", usersByEmailRoutes)
+    .route("/api/v2/users-by-email", usersByEmailRoutes)
     .route("/api/v2/clients", clientRoutes)
     .route("/api/v2/tenants", tenantRoutes)
     .route("/api/v2/logs", logRoutes)